Mise à jour des permissions (visiteur)

+++ src/controleurs/ControleurListe.php
@@ -22,6 +22,7 @@ class ControleurListe
     public function afficherToutesLesListes(){
         $listesAffichables = array();
         $listes = m\Liste::get();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
         if(isset($_SESSION['id'])){
             $userId = $_SESSION['id'];
             $user = m\User::where('id', '=', $userId)->first();
@@ -46,6 +47,7 @@ class ControleurListe
     public function afficherListe($lid) {
         $listeid = $lid;
         $liste = m\Liste::where('no', '=', $listeid)->first();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
         if(isset($_SESSION['id'])){
             $userId = $_SESSION['id'];
             $user = m\User::where('id', '=', $userId)->first();
@@ -64,14 +66,23 @@ class ControleurListe
      * Fonction permettant d'afficher les listes d'un utilisateur
      */
     public function afficherListeUtilisateur($userId) {
-        $userIdRegarder = $_SESSION['id'];
         $user = m\User::where('id', '=', $userId)->first();
-        $userRegarder = m\User::where('id', '=', $userIdRegarder)->first();
         $listes = $user->listes;
         $listesAffichables = array();
-        foreach($listes as $l){
-            if($l->peutAcceder($userRegarder)){
-                $listesAffichables[] = $l;
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $userIdRegarder = $_SESSION['id'];
+            $userRegarder = m\User::where('id', '=', $userIdRegarder)->first();
+            foreach($listes as $l){
+                if($l->peutAcceder($userRegarder)){
+                    $listesAffichables[] = $l;
+                }
+            }
+        } else {
+            foreach($listes as $l){
+                if($l->publique == 1){
+                    $listesAffichables[] = $l;
+                }
             }
         }
         $viewName = "Listes de $user->name";
@@ -79,19 +90,29 @@ class ControleurListe
     }
 
     public function afficherListeUtilisateurActuel(){
-        $this->afficherListeUtilisateur($_SESSION['id']);
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $this->afficherListeUtilisateur($_SESSION['id']);
+        } else {
+            (new v\ErreurView("Vous n'êtes pas connecté"))->renderFinal();
+        }
     }
 
     /**
      * Fonction permettant d'afficher le créateur liste
      */
     public function afficherCreateurListe() {
-        (new v\CreateListeView())->renderFinal();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            (new v\CreateListeView())->renderFinal();
+        } else {
+            (new v\ErreurView("Vous n'êtes pas connecté"))->renderFinal();
+        }
     }
 
     public function creerListe() {
-        //VERIFICATION DE LA CONNEXION
-        if(!empty($_SESSION)){
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
             $l = new m\Liste();
             $l->titre = filter_var($_POST['titre'], FILTER_SANITIZE_STRING);
             $l->description = filter_var($_POST['description'], FILTER_SANITIZE_STRING);
@@ -100,6 +121,8 @@ class ControleurListe
 
             $app = \Slim\Slim::getInstance();
             $app->redirect($app->urlFor('afficherListe', array('lid' => $l->no)));
+        } else {
+            (new v\ErreurView("Vous n'êtes pas connecté"))->renderFinal();
         }
     }
 
@@ -107,12 +130,14 @@ class ControleurListe
         $i = m\Item::where('id', '=', $id)->first();
         $liste = $i->liste;
         $liste_id = $i->liste_id;
-        $user = m\User::where('id', '=', $_SESSION['id'])->first();
-        //PROTECTION PERMISSIONS
-        if($liste->user_id == $_SESSION['id'] || $user->estParticipant($liste)){
-            $i->delete();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $user = m\User::where('id', '=', $_SESSION['id'])->first();
+            //PROTECTION PERMISSIONS
+            if($liste->user_id == $_SESSION['id'] || $user->estParticipant($liste)){
+                $i->delete();
+            }
         }
-
         $app = \Slim\Slim::getInstance();
         $app->redirect($app->urlFor('afficherListe', array('lid' => $liste_id)));
     }
@@ -121,14 +146,18 @@ class ControleurListe
         $user = m\User::where('email', '=', filter_var($_POST['email'], FILTER_SANITIZE_EMAIL))->first();
         $liste_id = filter_var($_POST['liste_id'], FILTER_SANITIZE_NUMBER_INT);
 
-        $l = m\Liste::where("no", "=", $liste_id)->first();
-        //PROTECTION PERMISSIONS
-        if ($l->user_id == $_SESSION["id"]) {
-            $p = new m\Participation();
-            $p->liste_id = $liste_id;
-            $p->user_id = $user->id;
-            $p->save();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $l = m\Liste::where("no", "=", $liste_id)->first();
+            //PROTECTION PERMISSIONS
+            if ($l->user_id == $_SESSION["id"]) {
+                $p = new m\Participation();
+                $p->liste_id = $liste_id;
+                $p->user_id = $user->id;
+                $p->save();
+            }
         }
+
         $app = \Slim\Slim::getInstance();
         $app->redirect($app->urlFor('afficherListe', ['lid' => $liste_id]));
     }
@@ -139,12 +168,15 @@ class ControleurListe
      * @param $user_id
      */
     public function supprimerParticipant($lid, $uid){
-        $liste_id = filter_var($lid, FILTER_SANITIZE_NUMBER_INT);
-        $user_id = filter_var($uid, FILTER_SANITIZE_NUMBER_INT);
-        $l = m\Liste::where("no", "=", $liste_id)->first();
-        //PROTECTION PERMISSIONS
-        if ($l->user_id == $_SESSION["id"] || $user_id == $_SESSION["id"]) {
-            m\Participation::where([['liste_id', '=', $liste_id], ['user_id', '=', $user_id]])->delete();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $liste_id = filter_var($lid, FILTER_SANITIZE_NUMBER_INT);
+            $user_id = filter_var($uid, FILTER_SANITIZE_NUMBER_INT);
+            $l = m\Liste::where("no", "=", $liste_id)->first();
+            //PROTECTION PERMISSIONS
+            if ($l->user_id == $_SESSION["id"] || $user_id == $_SESSION["id"]) {
+                m\Participation::where([['liste_id', '=', $liste_id], ['user_id', '=', $user_id]])->delete();
+            }
         }
         $app = \Slim\Slim::getInstance();
         $app->redirect($app->urlFor('afficherListe', array('lid' => $liste_id)));
@@ -153,14 +185,18 @@ class ControleurListe
     public function renommerUneListe(){
         //On recupere l'id de la liste
         $liste_id = filter_var($_POST['liste_id'], FILTER_SANITIZE_NUMBER_INT);
-        //On recupere la liste assosiciee
-        $liste = m\Liste::where('no', '=', $liste_id)->first();
-        //On recupere le nouveau titre
-        //PROTECTION PERMISSIONS
-        if ($liste->user_id == $_SESSION["id"]) {
-            $liste->titre = filter_var($_POST['titre'], FILTER_SANITIZE_STRING);
-            $liste->description = filter_var($_POST['description'], FILTER_SANITIZE_STRING);
-            $liste->save();
+
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            //On recupere la liste assosiciee
+            $liste = m\Liste::where('no', '=', $liste_id)->first();
+            //On recupere le nouveau titre
+            //PROTECTION PERMISSIONS
+            if ($liste->user_id == $_SESSION["id"]) {
+                $liste->titre = filter_var($_POST['titre'], FILTER_SANITIZE_STRING);
+                $liste->description = filter_var($_POST['description'], FILTER_SANITIZE_STRING);
+                $liste->save();
+            }
         }
 
         $app = \Slim\Slim::getInstance();
@@ -168,13 +204,16 @@ class ControleurListe
     }
 
     public function supprimerUneListe(){
-        //On recupere l'id de la liste
-        $liste_id = filter_var($_POST['liste_id'], FILTER_SANITIZE_NUMBER_INT);
-        //On recupere la liste assosiciee
-        $liste = m\Liste::where('no', '=', $liste_id)->first();
-        //PROTECTION PERMISSIONS
-        if ($liste->user_id == $_SESSION["id"]) {
-            $liste->delete();
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            //On recupere l'id de la liste
+            $liste_id = filter_var($_POST['liste_id'], FILTER_SANITIZE_NUMBER_INT);
+            //On recupere la liste assosiciee
+            $liste = m\Liste::where('no', '=', $liste_id)->first();
+            //PROTECTION PERMISSIONS
+            if ($liste->user_id == $_SESSION["id"]) {
+                $liste->delete();
+            }
         }
 
         $app = \Slim\Slim::getInstance();
@@ -184,52 +223,56 @@ class ControleurListe
     public function ajouterItem(){
         //VERIFICATION DES PERMISSIONS DE L'UTILISATEUR
         $liste_id = filter_var($_POST['liste_id'], FILTER_SANITIZE_NUMBER_INT);
-        $liste = m\Liste::where('no', '=', $liste_id)->first();
 
-        $user = m\User::where('id', '=', $_SESSION['id'])->first();
-        //PROTECTION PERMISSIONS
-        if($liste->user_id == $_SESSION['id'] || $user->estParticipant($liste)){
+        //SI L'UTILISATEUR EST CONNECTÉ AU SITE
+        if(isset($_SESSION['id'])){
+            $liste = m\Liste::where('no', '=', $liste_id)->first();
 
-            //FICHIER DE L'IMAGE
-            $target_dir = "web/uploads/";
-            $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
-            $uploadOk = 1;
-            $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+            $user = m\User::where('id', '=', $_SESSION['id'])->first();
+            //PROTECTION PERMISSIONS
+            if($liste->user_id == $_SESSION['id'] || $user->estParticipant($liste)){
 
-            //Verification du type de fichier (image)
-            if(isset($_POST["submit"])) {
-                $check = getimagesize($_FILES["fileToUpload"]["tmp_name"]);
-                if($check !== false) {
-                    $uploadOk = 1;
-                } else {
+                //FICHIER DE L'IMAGE
+                $target_dir = "web/uploads/";
+                $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
+                $uploadOk = 1;
+                $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+
+                //Verification du type de fichier (image)
+                if(isset($_POST["submit"])) {
+                    $check = getimagesize($_FILES["fileToUpload"]["tmp_name"]);
+                    if($check !== false) {
+                        $uploadOk = 1;
+                    } else {
+                        $uploadOk = 0;
+                    }
+                }
+                //Vérification de la taille du fichier
+                if ($_FILES["fileToUpload"]["size"] > 500000) {
                     $uploadOk = 0;
                 }
-            }
-            //Vérification de la taille du fichier
-            if ($_FILES["fileToUpload"]["size"] > 500000) {
-                $uploadOk = 0;
-            }
-            //Vérification du format du fichier
-            if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
-                $uploadOk = 0;
-            }
-            //Si toutes les vérifications sont bonnes
-            if ($uploadOk != 0) {
-                //AJOUT DE L'ITEM POUR OBTENIR L'ID
-                $i = new m\Item();
-                $i->nom = filter_var($_POST['nom'], FILTER_SANITIZE_STRING);
-                $i->descr = filter_var($_POST['description'], FILTER_SANITIZE_STRING);
-                $i->tarif = filter_var($_POST['prix'], FILTER_SANITIZE_NUMBER_FLOAT);
-                $i->liste_id = $liste_id;
-                $i->save();
-
-                //Verification de l'upload de l'image
-                $target_file = $target_dir . "item" . $i->id . "." . $imageFileType;
-                if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
-                    $i->img = "/" . $target_file;
+                //Vérification du format du fichier
+                if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
+                    $uploadOk = 0;
+                }
+                //Si toutes les vérifications sont bonnes
+                if ($uploadOk != 0) {
+                    //AJOUT DE L'ITEM POUR OBTENIR L'ID
+                    $i = new m\Item();
+                    $i->nom = filter_var($_POST['nom'], FILTER_SANITIZE_STRING);
+                    $i->descr = filter_var($_POST['description'], FILTER_SANITIZE_STRING);
+                    $i->tarif = filter_var($_POST['prix'], FILTER_SANITIZE_NUMBER_FLOAT);
+                    $i->liste_id = $liste_id;
                     $i->save();
-                } else {
-                    $i->delete();
+
+                    //Verification de l'upload de l'image
+                    $target_file = $target_dir . "item" . $i->id . "." . $imageFileType;
+                    if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
+                        $i->img = "/" . $target_file;
+                        $i->save();
+                    } else {
+                        $i->delete();
+                    }
                 }
             }
         }
